requirements.txt
openai==2.29.0
rich==14.1.0
python-dotenv==1.1.1
pytest==8.4.2
requests==2.32.5